Remove initializer for Rails < 6.1
In solidusio/solidus_support#65 and solidusio/solidus_support#71
they changed how the engine is loaded using the Rails initializers.
This updates causes some issues when using Rails < 6.1 because
the initializers are totally different. The solution is to load
the extension outside of the initializer block until we use Rails 6.1
or newer version.

Ref: solidusio/solidus_support#73.